refactor(friends): switch to route params instead of store
