+++ libraries/Enums/types/friendsTabs.ts
@@ -1,0 +1,5 @@
+export enum FriendsTabs {
+  DEFAULT = '',
+  ADD = 'add',
+  REQUESTS = 'requests',
+}
