+++ locales/en-US.js
@@ -732,9 +732,9 @@ export default {
     not_found: "Hmm, we couldn't find a user at that address",
     invalid_id: 'Invalid account ID',
     request_sent: 'Friend request successfully sent!',
-    requests: 'Friend requests',
+    friend_requests: 'Friend requests',
     new_friend_request: 'New friend request from {user}',
-    incoming: 'Requests',
+    requests: 'Requests',
     received: 'Received',
     sent: 'Sent',
     no_friend_request: 'You have no new friend requests.',
