+++ store/friends/actions.test.ts.disable
@@ -276,41 +276,3 @@ describe('default functions', () => {
     expect(commit).toHaveBeenCalledWith('removeFriend', '0.0.0.0') // the second argument here can be seen in the huge chunk of line above this line
   })
 })
-
-describe('utility functions', () => {
-  test('module.exportForTesting.friendAccountToIncomingRequest', () => {
-    const friendAccount = {
-      accountId: 'Checking Account',
-      from: '.',
-      status: 429,
-      fromMailboxId: '12345',
-      toMailboxId: 'v4.0.0-rc.4',
-      to: './path/to/file',
-    }
-    const userInfo = null
-
-    const result = module.exportForTesting.friendAccountToIncomingRequest(
-      friendAccount,
-      userInfo,
-    )
-    expect(result.account).toEqual(friendAccount)
-  })
-
-  test('module.exportForTesting.friendAccountToOutgoingRequest', () => {
-    const friendAccount = {
-      accountId: 'Checking Account',
-      from: '.',
-      status: 429,
-      fromMailboxId: '12345',
-      toMailboxId: 'v4.0.0-rc.4',
-      to: './path/to/file',
-    }
-    const userInfo = null
-
-    const result = module.exportForTesting.friendAccountToOutgoingRequest(
-      friendAccount,
-      userInfo,
-    )
-    expect(result.account).toEqual(friendAccount)
-  })
-})
