+++ store/friends/mutations.test.ts
@@ -1363,90 +1363,3 @@ describe('mutations.removeFriend', () => {
     expect(result).toMatchSnapshot()
   })
 })
-
-describe('mutations.sortFriends', () => {
-  test('0', () => {
-    const object: any = [
-      [true, false, true, false],
-      [true, true, true, false],
-      [true, true, true, false],
-      [true, false, true, false],
-    ]
-    const result: any = mutations.sortFriends({ all: object }, [
-      'https://api.telegram.org/',
-      'https://api.telegram.org/',
-      'www.google.com',
-      'https://croplands.org/app/a/reset?token=',
-    ])
-    expect(result).toMatchSnapshot()
-  })
-
-  test('1', () => {
-    const object: any = [
-      [false, false, false, false],
-      [true, false, true, false],
-      [true, false, false, false],
-      [false, false, false, true],
-    ]
-    const result: any = mutations.sortFriends({ all: object }, [
-      'Www.GooGle.com',
-      'https://',
-      'www.google.com',
-      '.com',
-    ])
-    expect(result).toMatchSnapshot()
-  })
-
-  test('2', () => {
-    const object: any = [
-      [false, false, false, false],
-      [true, true, false, false],
-      [true, false, true, false],
-      [true, false, false, true],
-    ]
-    const result: any = mutations.sortFriends({ all: object }, [
-      'https://accounts.google.com/o/oauth2/revoke?token=%s',
-      'www.google.com',
-      'Www.GooGle.com',
-      'Www.GooGle.com',
-    ])
-    expect(result).toMatchSnapshot()
-  })
-
-  test('3', () => {
-    const object: any = [
-      [true, false, true, false],
-      [false, false, true, true],
-      [false, true, false, false],
-      [true, true, true, false],
-    ]
-    const result: any = mutations.sortFriends({ all: object }, [
-      'https://croplands.org/app/a/reset?token=',
-      'http://base.com',
-      'http://www.croplands.org/account/confirm?t=',
-      'http://base.com',
-    ])
-    expect(result).toMatchSnapshot()
-  })
-
-  test('4', () => {
-    const object: any = [
-      [false, false, false, true],
-      [true, false, true, true],
-      [false, true, true, false],
-      [true, true, false, false],
-    ]
-    const result: any = mutations.sortFriends({ all: object }, [
-      'https://',
-      'https://api.telegram.org/bot',
-      'https://croplands.org/app/a/confirm?t=',
-      'www.google.com',
-    ])
-    expect(result).toMatchSnapshot()
-  })
-
-  test('5', () => {
-    const result: any = mutations.sortFriends({ all: [] }, [])
-    expect(result).toMatchSnapshot()
-  })
-})
