+++ store/friends/mutations.ts
@@ -1,5 +1,4 @@
-import { Conversation } from '../textile/types'
-import { FriendsState, FriendsTabs } from './types'
+import { FriendsState } from './types'
 import { Friend, IncomingRequest, OutgoingRequest } from '~/types/ui/friends'
 
 const mutations = {
@@ -95,18 +94,6 @@ const mutations = {
   },
   removeFriend(state: FriendsState, accountAddress: string) {
     state.all = state.all.filter((fr) => fr.address !== accountAddress)
-  },
-  sortFriends(state: FriendsState, rConversations: Conversation) {
-    state.all = state.all
-      .map((user) => {
-        const conversation = rConversations[user.address]
-        user.lastUpdate = conversation?.lastUpdate || 0
-        return user
-      })
-      .sort((user1, user2) => user2.lastUpdate - user1.lastUpdate)
-  },
-  setActiveTab(state: FriendsState, tab: FriendsTabs) {
-    state.activeTab = tab
   },
 }
