+++ store/friends/state.ts
@@ -1,11 +1,10 @@
-import { FriendsState, FriendsTabs } from '~/store/friends/types'
+import { FriendsState } from '~/store/friends/types'
 
 const InitialFriendsState: FriendsState = {
   incomingRequests: [],
   outgoingRequests: [],
   all: [],
   activeConversation: undefined,
-  activeTab: FriendsTabs.Add,
 }
 
 export default () => InitialFriendsState
