+++ store/friends/types.ts
@@ -2,14 +2,7 @@ import { Friend, IncomingRequest, OutgoingRequest } from '~/types/ui/friends'
 import { FriendAccount } from '~/libraries/Solana/FriendsProgram/FriendsProgram.types'
 import { UserInfo } from '~/libraries/Solana/UsersProgram/UsersProgram'
 
-export enum FriendsTabs {
-  Friends = 'friends',
-  Add = 'add',
-  Requests = 'requests',
-}
-
 export interface FriendsState {
-  activeTab: FriendsTabs
   incomingRequests: IncomingRequest[]
   outgoingRequests: OutgoingRequest[]
   all: Friend[]
@@ -27,42 +20,4 @@ export enum FriendsError {
 
 export interface FriendRequestArguments {
   did: string
-}
-
-/**
- * Utility function that converts a FriendAccount
- * into an strongly typed IncomingRequest
- * @param friendAccount the FriendAccount data
- * @returns IncomingRequest object
- */
-export function friendAccountToIncomingRequest(
-  friendAccount: FriendAccount,
-  userInfo: UserInfo | null,
-): IncomingRequest {
-  return {
-    requestId: friendAccount.accountId,
-    from: friendAccount.from,
-    account: friendAccount,
-    pending: false,
-    userInfo,
-  }
-}
-
-/**
- * Utility function that converts a FriendAccount
- * into an strongly typed OutgoingRequest
- * @param friendAccount the FriendAccount data
- * @returns OutgoingRequest object
- */
-export function friendAccountToOutgoingRequest(
-  friendAccount: FriendAccount,
-  userInfo: UserInfo | null,
-): OutgoingRequest {
-  return {
-    requestId: friendAccount.accountId,
-    to: friendAccount.to,
-    account: friendAccount,
-    pending: false,
-    userInfo,
-  }
 }
